# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: À Chacun Une Belle Âme — API
  version: 1.0.0-mvp
  description: |
    API REST de la plateforme de rencontres sérieuses « À Chacun Une Belle Âme ».

    ## Règles transverses
    - Toutes les dates sont en UTC (RFC 3339).
    - Tous les montants sont exprimés en unité monétaire minimale (`amountMinor`) avec
      la devise et son exposant explicites. **XAF et XOF ont un exposant 0.**
    - Toute liste est paginée par curseur opaque. Aucun paramètre `offset` n'existe.
    - Chaque opération déclare sa politique d'autorisation dans `x-authorization`.
    - Les opérations marquées `x-mock: true` reposent, tant qu'aucun prestataire n'est
      contractualisé, sur une implémentation simulée. Voir `docs/MOCKS.md`.
  contact:
    name: Équipe technique
servers:
  - url: https://api.example.invalid/api/v1
    description: Production (hôte à définir — question Q12)
  - url: http://localhost:3000/api/v1
    description: Développement local
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      operationId: postAuthRegister
      tags: [auth]
      summary: Inscription par numéro de téléphone
      description: |
        L'âge est **calculé côté serveur** à partir de `birthDate`. Une date de naissance
        correspondant à un âge inférieur au minimum légal entraîne un refus définitif
        (`AUTH_UNDERAGE`) et l'enregistrement d'une empreinte interdisant la ré-inscription.
      security: []
      x-authorization: PUBLIC
      x-rate-limit: 5/h par IP, 3/jour par numéro
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '202':
          description: Défi OTP émis
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpChallengeResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403':
          description: Inscription refusée (âge minimum non respecté ou identité bloquée)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Numéro déjà utilisé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /auth/otp/verify:
    post:
      operationId: postAuthOtpVerify
      tags: [auth]
      summary: Valider un code OTP et obtenir les jetons
      description: |
        Transport du refresh token selon `device.type` (story F-02) :
        - `WEB` : posé en cookie `httpOnly` (`__Secure-acuba_rt`, `Path=/api/v1/auth`), **absent
          du corps**. L'en-tête `Origin` doit figurer dans `CORS_ALLOWED_ORIGINS`, sinon 403
          `AUTH_FORBIDDEN` — contrôlé avant de consommer le code.
        - autre ou absent (application mobile) : rendu dans le corps, aucun cookie.
      security: []
      x-authorization: PUBLIC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, code]
              properties:
                challengeId: { type: string }
                code: { type: string, pattern: '^[0-9]{6}$' }
                device: { $ref: '#/components/schemas/DeviceInfo' }
      responses:
        '200':
          description: Authentifié
          headers:
            Set-Cookie:
              description: Navigateur uniquement — refresh token en cookie `httpOnly`
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403':
          description: Navigateur dont l'origine n'est pas déclarée (`AUTH_FORBIDDEN`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Code invalide, expiré ou nombre d'essais dépassé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/refresh:
    post:
      operationId: postAuthRefresh
      tags: [auth]
      summary: Rotation du refresh token
      description: |
        Le refresh token est à usage unique. La présentation d'un token déjà consommé est
        traitée comme un vol : toute la famille de sessions est révoquée (`AUTH_TOKEN_REUSED`).

        Deux transports (story F-02) :
        - `refreshToken` dans le corps (application mobile) : réponse dans le corps, aucun cookie.
        - corps `{}` et cookie `httpOnly` (navigateur) : `Origin` obligatoire et déclarée dans
          `CORS_ALLOWED_ORIGINS` (protection CSRF), cookie pivoté, `refreshToken` **absent** du
          corps. En cas d'échec, le cookie est effacé.
      security: []
      x-authorization: PUBLIC (le refresh token fait foi)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string, description: 'Absent pour un navigateur' }
      responses:
        '200':
          description: Nouveaux jetons
          headers:
            Set-Cookie:
              description: Transport par cookie uniquement — cookie pivoté, ou effacé en cas d'échec
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Transport par cookie depuis une origine absente ou non déclarée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/logout-all:
    post:
      operationId: postAuthLogoutAll
      tags: [auth]
      summary: Déconnecter tous les appareils
      x-authorization: AUTH
      responses:
        '204': { description: 'Toutes les sessions sont révoquées' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/sessions:
    get:
      operationId: getAuthSessions
      tags: [auth]
      summary: Lister les sessions et appareils actifs
      x-authorization: AUTH
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Sessions actives
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CursorPage'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Compte courant, statuts et étape d'onboarding suivante
      x-authorization: AUTH_PENDING
      responses:
        '200':
          description: Compte
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ─────────────────────────── VERIFICATION ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton d'accès de courte durée (15 minutes).
  parameters:
    Limit:
      name: limit
      in: query
      description: Nombre d'éléments (1 à 50, défaut 20).
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
    Cursor:
      name: cursor
      in: query
      description: Curseur opaque renvoyé par la page précédente.
      schema: { type: string }
...
  responses:
    ValidationError:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Authentification requise ou jeton invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: Trop de requêtes
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, timestamp]
          properties:
            code:
              type: string
              description: Code métier stable, jamais renommé.
              example: MSG_NO_MATCH
            message:
              type: string
              description: Message destiné à l'utilisateur, en français.
              example: Vous ne pouvez écrire qu'après un accord mutuel.
            details:
              oneOf:
                - type: object
                  additionalProperties: true
                - type: 'null'
            requestId: { type: string }
            timestamp: { type: string, format: date-time }
    CursorPage:
      type: object
      required: [items, hasMore]
      properties:
        items: { type: array, items: {} }
        nextCursor:
          oneOf:
            - type: string
            - type: 'null'
        hasMore: { type: boolean }
    RegisterRequest:
      type: object
      required: [phoneE164, birthDate, gender, consents]
      properties:
        phoneE164: { type: string, example: '+237690000000' }
        birthDate: { type: string, format: date, example: '1992-04-17' }
        gender: { type: string, enum: [FEMALE, MALE] }
        firstName: { type: string, maxLength: 60 }
        inviteCode: { type: string }
        consents:
          type: array
          minItems: 3
          items:
            type: object
            required: [type, documentVersion, granted]
            properties:
              type:
                type: string
                enum:
                  [
                    TERMS_OF_SERVICE,
                    PRIVACY_POLICY,
                    CODE_OF_CONDUCT,
                    KYC_PROCESSING,
                    MARKETING_COMMUNICATIONS,
                  ]
              documentVersion: { type: string }
              granted: { type: boolean }
        device: { $ref: '#/components/schemas/DeviceInfo' }
    DeviceInfo:
      type: object
      properties:
        type: { type: string, enum: [ANDROID, IOS, WEB] }
        model: { type: string }
        osVersion: { type: string }
        appVersion: { type: string }
        pushToken: { type: string }
    OtpChallengeResponse:
      type: object
      properties:
        challengeId: { type: string }
        expiresAt: { type: string, format: date-time }
        resendAvailableAt: { type: string, format: date-time }
        testMode:
          type: boolean
          description: Vrai lorsque l'envoi de SMS est simulé (le code est alors journalisé en développement).
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string, description: 'Absent pour un navigateur : voir le cookie' }
        expiresIn: { type: integer, example: 900 }
        user: { $ref: '#/components/schemas/Me' }
    Me:
      type: object
      properties:
        id: { type: string }
        accountStatus:
          type: string
          enum:
            [
              PENDING_OTP,
              ACTIVE,
              PAUSED,
              RESTRICTED,
              SUSPENDED,
              BANNED,
              BLOCKED_UNDERAGE,
              PENDING_DELETION,
              DELETED,
            ]
        verificationStatus:
          type: string
          enum:
            [NOT_STARTED, PENDING, IN_REVIEW, VERIFIED, REJECTED, ADDITIONAL_REQUIRED, SUSPENDED]
        profileCompletionRate: { type: integer, minimum: 0, maximum: 100 }
        isPremium: { type: boolean }
        nextOnboardingStep:
          type: string
          enum: [VERIFY_PHONE, VERIFY_IDENTITY, COMPLETE_PROFILE, ADD_PHOTOS, SET_PREFERENCES, DONE]
    Session:
      type: object
      properties:
        id: { type: string }
        deviceLabel: { type: string }
        current: { type: boolean }
        lastUsedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...